Move active-window code to vendors folder
