+++ lib/main.js
@@ -4,7 +4,7 @@ const realSessions = require('./real-session-data')
 let sessionsArray = realSessions.realSessions
 
 const menubar = require('menubar')
-const monitor = require('./active-window')
+const monitor = require('./vendors/active-window')
 
 // TODO resize this for production
 const mb = menubar({
