+++ lib/vendors/active-window/.npmignore
