+++ lib/vendors/active-window/configs.json
