+++ lib/vendors/active-window/index.js
