+++ lib/vendors/active-window/package.json
