+++ lib/vendors/active-window/scripts/linux.sh
