+++ lib/vendors/active-window/scripts/mac.scpt
