+++ lib/vendors/active-window/scripts/mac.sh
