+++ lib/vendors/active-window/scripts/windows.ps1
